end of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { environment } from 'src/environments/environment';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -18,7 +19,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 3;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -40,8 +41,12 @@ export class MessagesComponent implements OnInit {
     }
 
     deleteMessage(id: number) {
-      return this.messageService.deleteMessage(id).subscribe(() => {
-        this.messages.splice(this.messages.findIndex(m => m.id == id),1);
+      this.confirmService.confirm('confirm delete message', 'this cannot be undone').subscribe( result => {
+        if (result) {
+          return this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id == id),1);
+          })
+        }
       })
     }
 }
